项目任务优化调整，加回写项目，阶段，版本进度

+++ alm/src/main/java/com/dili/alm/controller/ProjectChangeController.java
@@ -61,7 +61,7 @@ public class ProjectChangeController {
 	@RequestMapping(value = "/loadTask", method = RequestMethod.GET)
 	@ResponseBody
 	public Object loadTask(Task task) throws Exception {
-		task.setStatus(ProjectState.COMPLETED.getValue().byteValue());
+		task.setStatus(ProjectState.COMPLETED.getValue().intValue());
 		return taskService.listEasyuiPageByExample(task,true).toString();
 	}
